+++ src/article/editor/styles/_context-section.css
@@ -3,10 +3,6 @@
   flex-direction: column;
 }
 
-.sc-context-section > .sc-navigator {
-  flex: 0 0 40px;
-}
-
 .sc-context-section > .se-context-content {
   flex: 1 1 0;
   overflow: auto;
