+++ src/article/editor/styles/_index.css
@@ -26,7 +26,6 @@
 @import './_fig-preview.css';
 @import './_table-fig-preview.css';
 @import './_fn-preview.css';
-@import './_navigator.css';
 @import './_translations.css';
 @import './_context-section.css';
 @import './_edit-reference-workflow.css';
