+++ src/article/shared/TexMathComponent.js
@@ -1,24 +1,7 @@
 import { NodeComponent } from '../../kit'
 import katex from 'katex'
 
-export default class InlineFormulaComponent extends NodeComponent {
-  render ($$) {
-    const node = this.props.node
-    const texMath = node.find('tex-math')
-    const el = $$('span').addClass('sc-inline-formula')
-    el.append(
-      $$(TexMathComponent, {
-        node: texMath
-      }).ref('math')
-    )
-    if (this.props.isolatedNodeState) {
-      el.addClass('sm-' + this.props.isolatedNodeState)
-    }
-    return el
-  }
-}
-
-class TexMathComponent extends NodeComponent {
+export default class TexMathComponent extends NodeComponent {
   render ($$) {
     const node = this.props.node
     const texMath = node.textContent
